LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/animations/transition-timing-function-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: Unexpected animation count expected 5 but got 6

PASS Ensure that transition easing functions are properly applied.
